requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow>=2.2.0
h5py
imgaug
IPython[all]
